Register dapp modal refactoring
